backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# OpenAI SDK
openai>=1.58.1

# Vector database
qdrant-client>=1.7.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.8.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0

# Text processing
markdown>=3.5.0
